+++ pkg/clog/mode.go
